+++ .github/workflows/miri.yaml
@@ -32,6 +32,7 @@ on:
       - arrow-schema/**
       - arrow-select/**
       - arrow-ipc/**
+      - arrow-csv/**
       - .github/**
 
 jobs:
